requirements.txt
azure-functions
azure-storage-blob
azure-ai-documentintelligence
semantic-kernel
pydantic==2.9.0
python-dotenv>=1.0.0
azure-core>=1.30.0
openai>=1.3.5
aiohttp>=3.9.1
typing-extensions>=4.8.0